backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# PDF processing
PyMuPDF>=1.24.0
pypdf==5.1.0

# Vector search and embeddings
faiss-cpu>=1.9.0
sentence-transformers>=3.3.1
numpy>=2.0.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# PDF creation (alternative to PyMuPDF for custom PDFs)
fpdf2==2.8.2
Pillow==11.0.0

# AI/LLM
openai>=1.54.0

